refactor: move functions to helper, styled components to styles/
